// src/app/Models/project.ts
export class Project {
    ProjectId: number;
    ProjectName: string;
    TasksCount: number;
    Completed: number;
    StartDate: string;
    EndDate: string;
    Priority: number;
    ProjectManagerId: number;
    ProjectManagerFullName: string;
}

// projects: Project[] = [
//     {
//       ProjectId: 1,
//       ProjectName: 'Development Project',
//       TasksCount: 5,
//       Completed: 2,
//       StartDate: '09/01/2018',
//       EndDate: '10/01/2018',
//       Priority: 1,
//       ProjectManagerId: 3049571,
//       ProjectManagerFullName: 'Uthaya Kumar Ganesan'
//     },
//     {
//       ProjectId: 2,
//       ProjectName: 'Testing Project',
//       TasksCount: 8,
//       Completed: 7,
//       StartDate: '09/05/2018',
//       EndDate: '10/10/2018',
//       Priority: 2,
//       ProjectManagerId: 1234567,
//       ProjectManagerFullName: 'Mani Krishna'
//     },
//     {
//       ProjectId: 3,
//       ProjectName: 'Enhancement Project',
//       TasksCount: 18,
//       Completed: 9,
//       StartDate: '09/01/2017',
//       EndDate: '10/01/2017',
//       Priority: 3,
//       ProjectManagerId: 9003483,
//       ProjectManagerFullName: 'Sunil Sinha Venkatesh'
//     }
//   ];